Port to QRandomGenerator
This sadly changes the output of a given "numbered game" so hopefully people don't have written down "nice mahjongg numbers"

Increase Qt dependency to 5.10 (the one that has QRandomGenerator) and KF5 one
to the first that depends on Qt 5.10

+++ src/gameview.h
@@ -406,7 +406,7 @@ private:
                    int yCounter, int z, int & zCount);
 
     unsigned short m_cheatsUsed;
-    long m_gameNumber;
+    quint32 m_gameNumber;
     bool m_gamePaused;
     bool m_match;
     bool m_gameGenerated;
